requirements.txt
requests
bs4
pillow
fpdf
PyPDF2